Set BlockChain on Backend just after construction
Previously this was being set on Start, which meant blockchain remained
nil for an indeterminate period of time. This fixes the panic that were
previously occurring in the cmd/autonity tests but they are still
failing because now due to the initing certain console output does not
match the expectations.