Updated TestCustomGenesis to use generated genesis
The tests are still failing because now the contract abi fails to be
retrieved, because the autonity contract is trying to retrieve it from
the database, this will be fixed by the work in
hardcode_deployer_address. And since this also depends on the setting
difficulty to 1 fix on develop we need to wait for
hardcode_deployer_address to be merged before this branch will be ready.